+++ packages/node-firebird-driver/src/lib/impl/statement.ts
@@ -104,8 +104,12 @@ export abstract class AbstractStatement implements Statement {
 		return resultSet;
 	}
 
+	get isValid(): boolean {
+		return !!this.attachment;
+	}
+
 	private check() {
-		if (!this.attachment)
+		if (!this.isValid)
 			throw new Error('Statement is already disposed.');
 	}