+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### driver, driver-native
 
 * Feature [#115](https://github.com/asfernandes/node-firebird-drivers/issues/115) - Add method Statement.getExecPathText.
+* Feature [#119](https://github.com/asfernandes/node-firebird-drivers/pull/119) - Implementation of isValid property.
 
 ### native-api